src/inventory_track/requirements.txt
Django>=5.1.5
mysqlclient>=2.1.1
pytz>=2023.3
numpy
pandas
matplotlib
requests
urllib3
python-dotenv